詳細画面の基盤を作成（これまでContentViewないでListの行のデザインや処理を行っていたが外部ファイルに分離させてさらにTodoをタップするとその詳細画面に遷移するようになっている）
